add deposit via native transfer
